Reader: refactor read_list
In order not to manually use runes and move the tokenizer in the
`read_list` proc, I had to re-write `read_form` as `read_token`, having
it accept a second Token parameter, and add these in two places.
I feel the `read_list` code is now a lot cleaner, and no complexity has
been added elsewhere.