Reader: store Symbols and Keywords as values
Originally these were stored as slices of the input buffer, which led to
them being overwritten, causing subtle bugs.
Now they are cloned into working memory, and can be stored indefinitely.

This change does not break any previous steps. In fact, it allowed me to
remove the symbol cloning from step 3 (and 4), as it was an ad-hoc
solution to a problem that originated in the reader.