odin: fix TCO bug in `do`
This somehow managed to slip through the cracks, but if I had something
like this:

(do stuff
    val)

Where `val` is a list, the list would get evaluated. This is a holdover
from the pre-TCO implementation of `do`, which I for some reason never
updated, and which passed all the tests (even though the instructions
state how it needs to be changed).